Move stat handler to subfolder
